Combined Parameter Data

Contains all combined parameter data for the Yolo Bypass Mass Balance Study. The calculations for this dataset are in the YB_Mass_Balance/Concentrations/Calculate_CombinedParams.R script file. The combined parameters include:

THg and MeHg Concentrations on Solids

Provide estimates of the amount of total mercury or methylmercury bound to the suspended sediment. An increasing or decreasing trend in these parameters indicate either an enrichment or reduction of Hg or MeHg on suspended particles.

THg and MeHg Partitioning Coefficients (Kd)

An increasing trend in these parameters indicates greater partitioning of Hg or MeHg to the solid-bound phase, and a decreasing trend indicates greater partitioning to the filtered or dissolved phase.

Percent MeHg Concentration of the THg Concentration

The percent of THg that is MeHg. An increasing trend indicates more of the total mercury is methylmercury. Values for all three fractions (filtered, particulate, and unfiltered) are provided.

THg or MeHg normalized by OC

Either THg or MeHg concentrations divided by organic carbon concentrations. Values for all three fractions (filtered, particulate, and unfiltered) are provided.

TOC or POC Concentrations on solids

Either TOC or POC concentrations divided by the TSS concentration multiplied by 1,000.

Total Aluminum Concentration on Solids

The total aluminum concentration divided by the TSS concentration multiplied by 1,000.

comb_param_calc

Format

data frame with 2,477 rows and 6 columns

StationName

The station name

SampleDate

The sampling date

CollectionTime

The collection time in PST

Parameter

The name of the calculated parameter

Value

Calculated value

Units

The units of the value