Convert StationName Variable to a Factor

Converts the StationName variable from a character variable to a factor variable to apply a consistent order in plots of the data. The StationName variable contains a standardized set of longer station names.

conv_fact_long_sta_names(df)

Arguments

df
The dataframe to apply the function to. Dataframe must have a character variable named StationName. Each observation under the StationName variable must have one of the following values:

Cache Slough near Ryer Island

CCSB- Low Flow Channel

CCSB Overflow Weir- North

CCSB Overflow Weir- South

Fremont Weir- East Side

Fremont Weir- Middle

Fremont Weir- West Side

Knights Landing Ridge Cut

Liberty Cut below Stairsteps

Miner Slough near Sac River

Prospect Slough

Putah Creek at Mace Blvd

Sac River above the Sacramento Weir

Shag Slough below Stairsteps

Toe Drain at 1/2 Lisbon

Toe Drain at County Road 22

Toe Drain at Interstate 80

Toe Drain at Lisbon Weir

Value

A dataframe with the StationName variable converted to a factor