Flow and Mass Load Data from the CALFED 2008 report

Contains total daily inflow and inlet, outlet, and net methylmercury load data from the 2006 flood event in the Yolo Bypass as presented in the CALFED 2008 report. This dataset is used as a comparison to the MeHg loads from the Yolo Bypass Mass Balance Study.

loads_flow_cf

Format

data frame with 14 rows and 7 columns

SampleDate

The date of the total inflow and net load values

Year

The year of the flood

Analyte

The analytical parameter for the net load value

Inlet_Load

The total inlet load value in g/day

Outlet_Load

The total outlet load value in g/day

Net_Load

The net load value in g/day

Total_Inflow

The total daily inflow value in cubic feet per second