Field and Filter Blanks Collected in Association with conc_data

Contains all field and filter blank data collected for the Yolo Bypass Mass Balance Study. Field and filter blanks were collected each day water samples were collected to test for possible contamination from sampling techniques, sample bottles, or filtering equipment. These blanks were separated from the raw data files in the YB_Mass_Balance/Concentrations/Compile_and_Clean_Conc_Data_FloodEvents.R script file.

qa_field_blanks

Format

data frame with 460 rows and 15 columns

SampleCode

The unique identifier for each field or filter blank sample produced by Bryte Lab

StationName

The name of the blank sample, either Field Blank or Filter Blank

SampleDate

The collection date of the blank sample

CollectionTime

The collection time of the blank sample in PST

Analyte

The analytical parameter measured

Result

The analytical result of the blank sample. "< RL" or "< MDL" indicates that the result was less than the Reporting Limit or Method Detection Limit, respectively.

ResultQual

Result qualifier for the blank sample. See Qual Codes and Flags for more information.

RL

The analytical reporting limit of the blank sample

MDL

The analytical method detection limit of the blank sample

Units

The units of the result

LabComments

Comments provided by the analytical laboratory

MME_Comments

Comments provided by staff from the DWR Mercury Monitoring and Evaluation section

AmbSampConc

Either the concentration of the ambient or environmental sample collected at the same station and date as the field blank sample, or the average concentration of all ambient samples associated with the filter blank sample. This value is only provided for blanks that had a detected value.

Blank_Amb_ratio

The ratio of the blank sample concentration to the AmbSampConc value expressed as a percent. If this percentage was greater than 20%, all samples collected on the same day were flagged as having a Blank Detection ("BD") in Flag.

Flag

Flag for the sample. "BD" indicates that there was a detection in the blank sample. See Qual Codes and Flags for more information.