Qual Codes and Flags

Following is more information on the Qual Codes and Flags used in the concentration datasets in the openwaterhg package.

Qual Codes

J

Not valid, but the flaw was determined to not be detrimental by best professional judgment; result can be used but with caution

R

Not valid, and existing information indicates that the result should not be used in any analysis

Flags

BD

Detection in Blank Sample: concentration in the blank sample was greater than 20% of the concentration of the associated ambient or environmental sample(s).

DNQ

Detected Not Quantifiable: value is greater than the Method Detection Limit (MDL) but less than the Reporting Limit (RL). Considered to be an estimate.

FV

Field Variability: the Relative percent difference (RPD) for associated field duplicate was greater than the Measurement Quality Objective of 25% (30% for VSS, TOC and DOC). Only applies when one or more of the sample results are greater than 10 times the RL or MDL (for parameters with MDL's provided).

FGT

Value of the filtered sample was greater than its associated total sample.

NRS

Sample may not have been representative of waterbody. Reason explained in MME_Comments.