Open Water Hg Report Plot Theme

This custom theme is the standardized theme used in plots within the final report for the Open Water Mercury studies. It builds off of the theme_light() ggplot theme and changes the facet label text color to black.

theme_owhg(x_axis_v = FALSE)

Arguments

x_axis_v
Makes the tick labels for the x-axis vertical if x_axis_v = TRUE. Default is FALSE.

Value

A ggplot layer that applies a custom theme as described above.